pyproject.toml
[project]
name = "simplecrm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "celery>=5.6.3",
    "django>=6.0.3",
    "google-api-python-client>=2.192.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "lxml>=6.0.2",
    "redis>=7.4.0",
]
